snAPI/requirements.txt
django-widget-tweaks
djangorestframework
djangorestframework-simplejwt
drf-yasg
mariadb
mysqlclient
packaging
pillow
PyJWT
python-decouple
sqlparse
tzdata
django-cors-headers
